+++ logger/config.go
@@ -2,6 +2,7 @@ package logger
 
 import "github.com/skeletonkey/lib-core-go/config"
 
+//nolint:gochecknoglobals // cfg holds the configuration data, which should only be retrieved and not changed
 var cfg *Logger
 
 func getConfig() *Logger {