+++ notify/pushover/main.go
@@ -2,48 +2,47 @@
 package pushover
 
 import (
+	"context"
+	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
 	"net/url"
-
-	"github.com/skeletonkey/lib-core-go/logger"
+	"strings"
 )
 
 // Notify sends `msg` using the Pushover API
-func Notify(msg string) error {
+// No timeout is set for the context - the caller should set one if desired
+func Notify(ctx context.Context, msg string) error {
 	config := getConfig()
-	log := logger.Get()
 	requestUrl := fmt.Sprintf("%s/messages.json?token=%s&user=%s&message=%s",
 		config.URL, config.Token.Application, config.Token.Account, url.QueryEscape(msg))
-	log.Trace().Str("URL", requestUrl).Msg("notification URL")
-	if !config.Enabled {
-		log.Info().Msg("Pushover is disabled")
-	} else {
-		res, err := http.Post(requestUrl, "application/json", nil)
-		if err != nil {
-			log.Debug().Err(err).Str("URL", requestUrl).Msg("unable to post to url")
-			return err
-		}
-		body, err := io.ReadAll(res.Body)
-		defer func() {
-			err := res.Body.Close()
-			if err != nil {
-				log.Error().
-					Err(err).
-					Msg("unable to close response body")
-			}
-		}()
-		if err != nil {
-			log.Debug().Err(err).Interface("response", res).Msg("unable to read response body")
-			return err
-		}
-		if res.StatusCode != 200 {
-			log.Debug().Int("Status Code", res.StatusCode).Bytes("response body", body).Msg("non-200 response received")
-			return err
-		}
 
-		log.Trace().Bytes("response body", body).Msg("pushover response")
+	if !config.Enabled {
+		return DisabledError{}
+	}
+
+	req, err := http.NewRequestWithContext(ctx, "POST", requestUrl, nil)
+	if err != nil {
+		return err
+	}
+	res, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return err
+	}
+	defer func() { _ = res.Body.Close() }()
+	if res.StatusCode/100 != 2 {
+		body, err := io.ReadAll(res.Body)
+		if err != nil {
+			return fmt.Errorf("pushover API error (%d): unable to read response body", res.StatusCode)
+		}
+		var apiResp struct {
+			Errors []string `json:"errors"`
+		}
+		if err := json.Unmarshal(body, &apiResp); err == nil && len(apiResp.Errors) > 0 {
+			return fmt.Errorf("pushover API error (%d): %s", res.StatusCode, strings.Join(apiResp.Errors, "; "))
+		}
+		return fmt.Errorf("pushover API error (%d): unknown error", res.StatusCode)
 	}
 
 	return nil